+++ tests/LightVarDumperProviders/ProviderMaxStringLength.php
@@ -1,0 +1,21 @@
+<?php
+
+namespace Awesomite\VarDumper\LightVarDumperProviders;
+
+/**
+ * @internal
+ */
+class ProviderMaxStringLength implements \IteratorAggregate
+{
+    public function getIterator()
+    {
+        $result = array(
+            array(5, 'Hello world!', "string(12) 'Hello'...\n"),
+            array(2, 'Hello world!', "string(12) 'He'...\n"),
+            array(12, 'Hello world!', "string(12) 'Hello world!'\n"),
+            array(13, 'Hello world!', "string(12) 'Hello world!'\n"),
+        );
+
+        return new \ArrayIterator($result);
+    }
+}